Move attack strategies to a new module and reorder imports
